deleted qaqc_test, resolved some merge conflicts
