+++ Backend/src/db/__init__.py
@@ -1,5 +1,5 @@
 from .getters import get_all_nodes_from, get_all_collections, get_all_labels_from, get_ml_status
-from .setters import add_one_node_to, add_multiple_nodes_to, add_about_node_to_id, add_multiple_nodes_to_id, set_ml_status
+from .setters import add_one_node_to, add_about_node_to_id, add_multiple_nodes_to_id, set_ml_status
 from .modifiers import delete_collection, delete_ml_status
 
 __all__ = [
@@ -10,7 +10,7 @@ __all__ = [
     "delete_collection",
     "add_about_node_to_id",
     "add_multiple_nodes_to_id",
-    "get_all_label_from",
+    "get_all_labels_from",
     "delete_ml_status",
     "set_ml_status",
     "get_ml_status",
