+++ Backend/src/ml_lib/factory/getter.py
@@ -1,6 +1,6 @@
 # own
 from ..shared import IsbitClassifierModel
-from ..sub_models import QaqcTestModel, QaqcMainModel
+from ..sub_models import QaqcMainModel
 
 
 def get_instance(model_name: str) -> IsbitClassifierModel:
@@ -8,8 +8,6 @@ def get_instance(model_name: str) -> IsbitClassifierModel:
     Factory to get a ML instance of the requested model.
     """
     match model_name:
-        case "qaqc_test":
-            return QaqcTestModel()
         case "qaqc_main":
             return QaqcMainModel()
         case _:
