+++ Backend/src/ml_lib/sub_models/__init__.py
@@ -1,3 +1,2 @@
-from .qaqc_test_model import QaqcTestModel
 from .qaqc_main_model import QaqcMainModel
-__all__ = ["QaqcTestModel","QaqcMainModel"]
+__all__ = ["QaqcMainModel"]
